+++ docs/source/api_ref_training.rst
@@ -14,3 +14,13 @@ Utilities for working in a reduced precision setting.
     :nosignatures:
 
     get_quantizer_mode
+
+
+Miscellaneous
+-------------
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    set_seed
