+++ tests/torchtune/models/clip/test_clip_image_transform.py
@@ -15,7 +15,7 @@ from torchtune.models.clip._transforms import CLIPImageTransform
 from torchtune.models.clip.inference._transforms import (
     CLIPImageTransform as CLIPImageTransformInference,
 )
-from torchtune.utils.seed import set_seed
+from torchtune.training.seed import set_seed
 
 
 @pytest.fixture(autouse=True)
