+++ tests/torchtune/models/phi3/test_lora_phi3.py
@@ -16,7 +16,7 @@ from torchtune import utils
 from torchtune.models.phi3 import lora_phi3, phi3
 from torchtune.models.phi3._component_builders import lora_phi3_self_attention
 from torchtune.modules.peft import get_merged_lora_ckpt, LoRALinear
-from torchtune.utils.seed import set_seed
+from torchtune.training.seed import set_seed
 
 RANK = 4
 ALPHA = 1.0
