+++ tests/torchtune/models/qwen2/test_qwen2.py
@@ -8,7 +8,7 @@ import pytest
 import torch
 from tests.test_utils import fixed_init_model
 from torchtune.models.qwen2 import qwen2
-from torchtune.utils.seed import set_seed
+from torchtune.training.seed import set_seed
 
 EMBED_DIM = 128
 INTER_DIM = 256
