+++ tests/torchtune/modules/low_precision/test_nf4_linear.py
@@ -14,7 +14,7 @@ import pytest
 import torch
 from torchao.dtypes.nf4tensor import NF4Tensor
 from torchtune.modules.low_precision import FrozenNF4Linear
-from torchtune.utils.seed import set_seed
+from torchtune.training.seed import set_seed
 
 
 @pytest.fixture(autouse=True)
