+++ tests/torchtune/modules/model_fusion/test_fusion_layer.py
@@ -10,7 +10,7 @@ import torch
 from tests.test_utils import assert_expected, fixed_init_model
 from torch import nn
 from torchtune.modules.model_fusion import FusionLayer
-from torchtune.utils.seed import set_seed
+from torchtune.training.seed import set_seed
 
 
 @pytest.fixture(autouse=True)
