+++ tests/torchtune/modules/test_layernorm.py
@@ -11,7 +11,7 @@ import torch
 from tests.test_utils import assert_expected
 
 from torchtune.modules.layer_norm import Fp32LayerNorm
-from torchtune.utils.seed import set_seed
+from torchtune.training.seed import set_seed
 
 
 @pytest.fixture(autouse=True)
