+++ tests/torchtune/training/test_seed.py
@@ -11,7 +11,7 @@ import os
 import numpy as np
 import pytest
 import torch
-from torchtune.utils.seed import set_seed
+from torchtune.training.seed import set_seed
 
 
 class TestSeed:
