+++ torchtune/training/__init__.py
@@ -5,7 +5,9 @@
 # LICENSE file in the root directory of this source tree.
 
 from torchtune.training.quantization import get_quantizer_mode
+from torchtune.training.seed import set_seed
 
 __all__ = [
     "get_quantizer_mode",
+    "set_seed",
 ]
